TextInput width,height값 props로 수정
