--- src/components/TextField/CommonInput/Common.module.scss
@@ -1,21 +1,0 @@
-.inputWrapper {
-    width: 423px;
-    height: 88px;
-
-    .inputTitle {
-        padding-bottom: 12px;
-        line-height: 28px;
-    }
-
-    .inputContent {
-        width: 423px;
-        height: 48px;
-        border: 1px solid #dddddd;
-    }
-
-    .requiredSymbol {
-        color: red;
-        font-size: 16px;
-        margin-left: 4px;
-    }
-}
