+++ src/components/TextField/CommonInput/TextInput.module.scss
@@ -1,0 +1,42 @@
+.inputWrapper {
+    width: 423px;
+    height: 88px;
+
+    .inputTitle {
+        padding-bottom: 12px;
+        line-height: 28px;
+    }
+
+    .inputContent {
+        width: 423px;
+        height: 48px;
+        border: 1px solid #dddddd;
+    }
+
+    .requiredSymbol {
+        .inputWrapper {
+            // width: 423px;
+            height: 88px;
+
+            .inputTitle {
+                padding-bottom: 12px;
+                line-height: 28px;
+            }
+
+            .inputContent {
+                // width: 423px;
+                // height: 48px;
+                border: 1px solid #dddddd;
+            }
+
+            .requiredSymbol {
+                color: red;
+                font-size: 16px;
+                margin-left: 4px;
+            }
+        }
+        color: red;
+        font-size: 16px;
+        margin-left: 4px;
+    }
+}
